// src/lib/schemas/RegisterSchema.ts
import { z } from "zod";

export const registerSchema = z.object({
    name: z.string().min(3, { message: 'Name must be at least 3 charachters' }),
    email: z.string().email(),
    password: z.string().min(6, { message: 'Password must be at least 6 charachters'})
})

export type RegisterSchema = z.infer<typeof registerSchema>